Separate controller package from CLI command and nest scheduler and monitor in it
