+++ internal/controller/scheduler/completions.go
