+++ internal/controller/scheduler/scheduler_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 
 	"github.com/buildkite/agent-stack-k8s/v2/api"
-	"github.com/buildkite/agent-stack-k8s/v2/internal/monitor"
-	"github.com/buildkite/agent-stack-k8s/v2/internal/scheduler"
+	"github.com/buildkite/agent-stack-k8s/v2/internal/controller/monitor"
+	"github.com/buildkite/agent-stack-k8s/v2/internal/controller/scheduler"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap/zaptest"
